add app deploy command
