+++ tests/nativeapp/utils.py
@@ -18,6 +18,9 @@ NATIVEAPP_MANAGER_APP_PKG_DISTRIBUTION_IN_SF = (
 NATIVEAPP_MANAGER_IS_APP_PKG_DISTRIBUTION_SAME = (
     f"{NATIVEAPP_MANAGER}.verify_project_distribution"
 )
+NATIVEAPP_MANAGER_GET_EXISTING_APP_PKG_INFO = (
+    f"{NATIVEAPP_MANAGER}.get_existing_app_pkg_info"
+)
 
 TEARDOWN_PROCESSOR_GET_EXISTING_APP_INFO = f"{TEARDOWN_PROCESSOR}.get_existing_app_info"
 TEARDOWN_PROCESSOR_GET_EXISTING_APP_PKG_INFO = (
@@ -27,7 +30,6 @@ TEARDOWN_PROCESSOR_IS_CORRECT_OWNER = f"{TEARDOWN_MODULE}.ensure_correct_owner"
 TEARDOWN_PROCESSOR_DROP_GENERIC_OBJECT = f"{TEARDOWN_PROCESSOR}.drop_generic_object"
 
 RUN_PROCESSOR_GET_EXISTING_APP_INFO = f"{RUN_PROCESSOR}.get_existing_app_info"
-RUN_PROCESSOR_GET_EXISTING_APP_PKG_INFO = f"{RUN_PROCESSOR}.get_existing_app_pkg_info"
 
 FIND_VERSION_FROM_MANIFEST = f"{VERSION_MODULE}.find_version_info_in_manifest_file"
